backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17
python-dotenv>=1.0.1
pydantic>=2.10.0
sse-starlette>=2.2.0
rapidfuzz>=3.10.0
jieba>=0.42.1
opentimelineio>=0.17.0
httpx>=0.28.0
openai>=1.60.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
